% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_location.R
\name{catch_location}
\alias{catch_location}
\title{Match catch species with fishing coordinates and bathymetry}
\usage{
catch_location(dffish, dftracker, depthraster)
}
\arguments{
\item{dffish}{ifish postgresQL data frame containing fish species, length data, and boat or landing information.}

\item{dftracker}{ifish postgresQL data frame containing tracker information, including latitude, longitude, and date-time.}

\item{depthraster}{bathymetry raster file}
}
\value{
data frame that combines catch data from CODRS photographs and fishing coordinates
        from Spot Trace. Fishing ground was matched with photographs based on first_codrs_photo_date and date of fishing coordinates.
        Fishing trips with no corresponding fishing coordinates are grouped by boat name and assigned its mean latitude and mean longitude.
}
\description{
Match catch species with fishing coordinates and bathymetry
}
