% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create a grid in the Indonesian EEZ}
\usage{
create_grid(eez, type = "polygons")
}
\arguments{
\item{eez}{sf object for Indonesian EEZ shapefile}

\item{type}{type of grid. Default type='polygon'. Other options are 'centers'. Argument in function must be a string.}
}
\description{
Create a grid in the Indonesian EEZ
}
