% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_coordinates.R
\name{fishing_grounds}
\alias{fishing_grounds}
\title{Get fishing coordinates}
\usage{
fishing_grounds(df, depthraster)
}
\arguments{
\item{df}{the ifish dataframe; must contain tables: ifish_tracker, ifish_boat, ifish_findmespot}

\item{depthraster}{the depth raster file, for example, GEBCO file}
}
\value{
original data frame with fishing coordinates, depth, and secondary bathymetric information added.
}
\description{
This function uses Spot Trace data to determine fishing coordinates, interpolate depth, and secondary bathymetric information.
}
