% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_coordinates.R
\name{simple_distance}
\alias{simple_distance}
\title{Get distance}
\usage{
simple_distance(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{longitude of point 1}

\item{lat1}{latitude of point 1}

\item{lon2}{longitude of point 2}

\item{lat2}{latitude of point 2}
}
\value{
the distance between two points. If lat1, lon1, lat2, and lon2 were supplied as a vector, it will return a vector of distances.
}
\description{
This function is to calculate the distane between two coordinates based on how a crow flies.
}
