% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{species_biomass}
\alias{species_biomass}
\title{Generate table for the top species in the catch by weight}
\usage{
species_biomass(df2)
}
\arguments{
\item{df2}{ifish postgresQL data frame}
}
\description{
Generate table for the top species in the catch by weight
}
