% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{species_biomass_gear}
\alias{species_biomass_gear}
\title{Generate table for the top species in the catch by weight and fishing gear}
\usage{
species_biomass_gear(df2)
}
\arguments{
\item{df2}{ifish postgresQL data frame}
}
\description{
Generate table for the top species in the catch by weight and fishing gear
}
