% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{species_freq}
\alias{species_freq}
\title{Generate table of the top species in the catch by frequency}
\usage{
species_freq(df2)
}
\arguments{
\item{df2}{ifish postgresQL data frame}
}
\description{
Generate table of the top species in the catch by frequency
}
